ENH moved all ML adapters into a dedicated folder
