+++ src/MLAdapters/ScrollZoomHandler.ts
@@ -4,7 +4,7 @@
 
 import maplibregl from "maplibre-gl";
 import type { Map as MapMLGL } from "maplibre-gl";
-import { Map } from "./Map";
+import { Map } from "../Map";
 
 export class ScrollZoomHandler extends maplibregl.ScrollZoomHandler {
   constructor(map: Map | MapMLGL, triggerRenderFrame: () => void) {
