+++ src/MLAdapters/Style.ts
@@ -4,7 +4,7 @@
 
 import maplibregl from "maplibre-gl";
 import type { Map as MapMLGL, StyleOptions } from "maplibre-gl";
-import { Map } from "./Map";
+import { Map } from "../Map";
 
 export class Style extends maplibregl.Style {
   constructor(map: Map, options: StyleOptions = {}) {
